+++ __tests__/functional/transaction-forging/htlc-lock.test.ts
@@ -1,4 +1,5 @@
-import { Crypto, Enums, Identities } from "@arkecosystem/crypto";
+import { Crypto, Enums, Identities, Utils } from "@arkecosystem/crypto";
+import got from "got";
 import { TransactionFactory } from "../../helpers/transaction-factory";
 import { secrets } from "../../utils/config/testnet/delegates.json";
 import * as support from "./__support__";
@@ -129,5 +130,80 @@ describe("Transaction Forging - HTLC Lock", () => {
         await expect(transaction).toBeAccepted();
         await support.snoozeForBlock(1);
         await expect(transaction.id).toBeForged();
+    });
+
+    it("should update delegates vote balance using locked balance when voting and unvoting delegates", async () => {
+        const newWalletPassphrase = "this is a new wallet passphrase";
+        // Initial Funds
+        const initialBalance = 100 * 1e8;
+        const initialFunds = TransactionFactory.transfer(
+            Identities.Address.fromPassphrase(newWalletPassphrase),
+            initialBalance,
+        )
+            .withPassphrase(secrets[0])
+            .createOne();
+
+        await expect(initialFunds).toBeAccepted();
+        await support.snoozeForBlock(1);
+        await expect(initialFunds.id).toBeForged();
+
+        const delegateToVote = Identities.PublicKey.fromPassphrase(secrets[9]);
+        const { body } = await got.get(`http://localhost:4003/api/v2/delegates/${delegateToVote}`);
+        const parsedBody = JSON.parse(body);
+        const initialDelegateVoteValance = Utils.BigNumber.make(parsedBody.data.votes);
+
+        // Submit a vote
+        const vote = TransactionFactory.vote(delegateToVote)
+            .withPassphrase(newWalletPassphrase)
+            .createOne();
+
+        await expect(vote).toBeAccepted();
+        await support.snoozeForBlock(1);
+        await expect(vote.id).toBeForged();
+
+        const expectedBalanceAfterVote = initialDelegateVoteValance.plus(initialBalance).minus(vote.fee);
+        await expect(delegateToVote).toHaveVoteBalance(expectedBalanceAfterVote.toString());
+
+        // Submit htlc lock transaction
+        const lockTransaction = TransactionFactory.htlcLock({
+            secretHash: "0f128d401958b1b30ad0d10406f47f9489321017b4614e6cb993fc63913c5454",
+            expiration: {
+                type: EpochTimestamp,
+                value: Crypto.Slots.getTime() + 1000,
+            },
+        })
+            .withPassphrase(newWalletPassphrase)
+            .createOne();
+
+        await expect(lockTransaction).toBeAccepted();
+        await support.snoozeForBlock(1);
+        await expect(lockTransaction.id).toBeForged();
+
+        const expectedBalanceAfterLock = expectedBalanceAfterVote.minus(lockTransaction.fee);
+        await expect(delegateToVote).toHaveVoteBalance(expectedBalanceAfterLock.toString());
+
+        // Unvote
+        const unvote = TransactionFactory.unvote(delegateToVote)
+            .withPassphrase(newWalletPassphrase)
+            .createOne();
+
+        await expect(unvote).toBeAccepted();
+        await support.snoozeForBlock(1);
+        await expect(unvote.id).toBeForged();
+
+        const expectedBalanceAfterUnvote = initialDelegateVoteValance;
+        await expect(delegateToVote).toHaveVoteBalance(expectedBalanceAfterUnvote.toString());
+
+        // Vote again
+        const voteAgain = TransactionFactory.vote(delegateToVote)
+            .withPassphrase(newWalletPassphrase)
+            .createOne();
+
+        await expect(voteAgain).toBeAccepted();
+        await support.snoozeForBlock(1);
+        await expect(voteAgain.id).toBeForged();
+
+        const expectedBalanceAfterVoteAgain = expectedBalanceAfterLock.minus(unvote.fee).minus(voteAgain.fee);
+        await expect(delegateToVote).toHaveVoteBalance(expectedBalanceAfterVoteAgain.toString());
     });
 });