+++ packages/core-jest-matchers/src/functional/index.ts
@@ -2,3 +2,4 @@ import "./accepted";
 import "./forged";
 import "./rejected";
 import "./unconfirmed";
+import "./vote-balance";